(see description) GameManager updated, still bug with 3 windows
Currently the only way to have it compile (while trying to transfer Irrlicht component init and more to GameManager) is to place these as static in the GameManager header file. Problem is that any class needing the components needs to include GameManager header file now thanks to this, which runs the component creation again, including new devices, which cause multi windows. I tried several things but in the end I had to set it back to this due to not being able to compile.

I need to do more research about C++ before I can find a fix

Robert