Added: grid functionality (offset,distance check)
Added: offset support to the randomObjectOnVertex function as a bool parameter so you can choose if you want to place the object within or with the offset of the playingfield.
Added: Distance check between chest and key. Now it will shuffle again if the distance is too short.

Change: This function makes use of a gameObjectList or a single gameobject. All gameObjects are made outside this function instead of in the function.

Still needs to be changed: Properties of all game objects (scale, rotation and amount).